move the game language setting from setting>Other to setting > Store > Steam
