+++ app/src/main/feature/settings/stores/StoresFragment.kt
@@ -25,6 +25,7 @@ import com.winlator.cmod.feature.stores.gog.service.GOGAuthManager
 import com.winlator.cmod.feature.stores.gog.service.GOGService
 import com.winlator.cmod.feature.stores.gog.ui.auth.GOGOAuthActivity
 import com.winlator.cmod.feature.stores.steam.SteamLoginActivity
+import com.winlator.cmod.feature.stores.steam.enums.Language
 import com.winlator.cmod.feature.stores.steam.service.SteamService
 import com.winlator.cmod.feature.stores.steam.utils.PrefManager
 import com.winlator.cmod.shared.android.DirectoryPickerDialog
@@ -150,6 +151,11 @@ class StoresFragment : Fragment() {
                         onPickSteamFolder = { pickFolder(PrefManager.steamDownloadFolder) { PrefManager.steamDownloadFolder = it } },
                         onPickEpicFolder = { pickFolder(PrefManager.epicDownloadFolder) { PrefManager.epicDownloadFolder = it } },
                         onPickGogFolder = { pickFolder(PrefManager.gogDownloadFolder) { PrefManager.gogDownloadFolder = it } },
+                        onContainerLanguageSelected = { index ->
+                            val langName = Language.containerLangForIndex(index)
+                            PrefManager.containerLanguage = langName
+                            refresh()
+                        },
                         bridge = (requireActivity() as? UnifiedActivity)?.settingsNavBridge,
                     )
                 }
@@ -165,6 +171,8 @@ class StoresFragment : Fragment() {
     // Helpers
     private fun refresh() {
         val ctx = context ?: return
+        val containerLanguageLabels = Language.displayLabels()
+        val containerLanguageIndex = Language.indexForContainerLang(PrefManager.containerLanguage)
         storeState =
             StoreState(
                 isSteamLoggedIn = SteamService.isLoggedIn,
@@ -178,6 +186,8 @@ class StoresFragment : Fragment() {
                 steamFolder = resolveUri(PrefManager.steamDownloadFolder, ctx),
                 epicFolder = resolveUri(PrefManager.epicDownloadFolder, ctx),
                 gogFolder = resolveUri(PrefManager.gogDownloadFolder, ctx),
+                containerLanguageLabels = containerLanguageLabels,
+                containerLanguageIndex = containerLanguageIndex,
             )
     }
